Added paper draft
At the last second yesterday I decided to switch from the more mathematical premise I was using - which was quickly turning into describing the relevance of differential geometry to gauge theory - to something which hopefully captures the physical essence of gauge theory better.  This draft is the culmination of my work under that new premise.